+++ src/main/java/javaclasses/library/event/book/BookUpdated.java
@@ -1,4 +1,4 @@
-package javaclasses.library.book;
+package javaclasses.library.event.book;
 
 import javaclasses.library.tinytype.BookDescription;
 import javaclasses.library.tinytype.BookId;
