+++ src/main/java/javaclasses/library/event/book/BookVO.java
@@ -1,4 +1,4 @@
-package javaclasses.library.book;
+package javaclasses.library.event.book;
 
 import javaclasses.library.tinytype.*;
 import javafx.scene.image.Image;
