+++ src/main/java/javaclasses/library/event/inventory/BookMarkedAsAvailable.java
@@ -1,22 +1,15 @@
-package javaclasses.library.inventory;
+package javaclasses.library.event.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.InventoryItemId;
-import javaclasses.library.tinytype.UserId;
 
-public class BookMarkedAsLost {
+public class BookMarkedAsAvailable {
     private final InventoryId inventoryId;
     private final InventoryItemId inventoryItemId;
-    private final UserId userId;
 
-    public BookMarkedAsLost(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
+    public BookMarkedAsAvailable(InventoryId inventoryId, InventoryItemId inventoryItemId) {
         this.inventoryId = inventoryId;
         this.inventoryItemId = inventoryItemId;
-        this.userId = userId;
-    }
-
-    public UserId getUserId() {
-        return userId;
     }
 
     public InventoryId getInventoryId() {
