+++ src/main/java/javaclasses/library/event/inventory/BookMarkedAsLost.java
@@ -1,15 +1,22 @@
-package javaclasses.library.inventory;
+package javaclasses.library.event.inventory;
 
-import javaclasses.library.tinytype.InventoryItemId;
 import javaclasses.library.tinytype.InventoryId;
+import javaclasses.library.tinytype.InventoryItemId;
+import javaclasses.library.tinytype.UserId;
 
-public class BookMarkedAsReturned {
+public class BookMarkedAsLost {
     private final InventoryId inventoryId;
     private final InventoryItemId inventoryItemId;
+    private final UserId userId;
 
-    public BookMarkedAsReturned(InventoryId inventoryId, InventoryItemId inventoryItemId) {
+    public BookMarkedAsLost(InventoryId inventoryId, InventoryItemId inventoryItemId, UserId userId) {
         this.inventoryId = inventoryId;
         this.inventoryItemId = inventoryItemId;
+        this.userId = userId;
+    }
+
+    public UserId getUserId() {
+        return userId;
     }
 
     public InventoryId getInventoryId() {
