+++ src/main/java/javaclasses/library/event/inventory/InventoryDecreased.java
@@ -1,22 +1,23 @@
-package javaclasses.library.inventory;
+package javaclasses.library.event.inventory;
 
 import javaclasses.library.tinytype.InventoryItemId;
 import javaclasses.library.tinytype.InventoryId;
 
-public class InventoryAppended {
+public class InventoryDecreased {
     private final InventoryItemId inventoryItemId;
     private final InventoryId inventoryId;
 
-    public InventoryAppended(InventoryItemId inventoryItemId, InventoryId inventoryId) {
+    public InventoryDecreased(InventoryItemId inventoryItemId, InventoryId inventoryId) {
         this.inventoryItemId = inventoryItemId;
         this.inventoryId = inventoryId;
-    }
 
-    public InventoryId getInventoryId() {
-        return inventoryId;
     }
 
     public InventoryItemId getInventoryItemId() {
         return inventoryItemId;
+    }
+
+    public InventoryId getInventoryId() {
+        return inventoryId;
     }
 }
