+++ src/main/java/javaclasses/library/event/inventory/LoanPeriodExtended.java
@@ -1,23 +1,22 @@
-package javaclasses.library.inventory;
+package javaclasses.library.event.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.UserId;
 
-public class LoanBecameOverdue {
+public class LoanPeriodExtended {
     private final UserId userId;
     private final InventoryId inventoryId;
 
-    public LoanBecameOverdue(UserId userId, InventoryId inventoryId) {
+    public LoanPeriodExtended(UserId userId, InventoryId inventoryId) {
         this.userId = userId;
         this.inventoryId = inventoryId;
-    }
-
-    public InventoryId getInventoryId() {
-        return inventoryId;
     }
 
     public UserId getUserId() {
         return userId;
     }
 
+    public InventoryId getInventoryId() {
+        return inventoryId;
+    }
 }
