+++ src/main/java/javaclasses/library/event/inventory/ReservationRemoved.java
@@ -1,13 +1,13 @@
-package javaclasses.library.inventory;
+package javaclasses.library.event.inventory;
 
 import javaclasses.library.tinytype.InventoryId;
 import javaclasses.library.tinytype.UserId;
 
-public class ReservationCanceled {
+public class ReservationRemoved {
     private final InventoryId inventoryId;
     private final UserId userId;
 
-    public ReservationCanceled(InventoryId inventoryId, UserId userId) {
+    public ReservationRemoved(InventoryId inventoryId, UserId userId) {
         this.inventoryId = inventoryId;
         this.userId = userId;
     }
